requirements.txt
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.9.4
dnspython>=2.6.0
scikit-learn>=1.4.0
pandas>=2.2.0
joblib>=1.3.0
rich>=13.7.0
pytest>=8.0.0
